requirements.txt
fastapi
uvicorn
pandas
python-dotenv
groq
duckdb
